Rename helper funding reducer
